escena de pick de cartas
ya manda bien a los jugadores, hay que pobrar las funcionalidades ahora

+++ Assets/Scripts/scriptable objects/cards.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "New Card", menuName = "Card")]
-public class cards :ScriptableObject
+public class cards : ScriptableObject
 {
     public string name;
     public float damage;
     public float cooldown;
-    public int bullets;
+    public float bullets;
 
     public void Print()
     {
